chore(EMS-2308): add/improve e2e test coverage for companies house summary lists
